Novas alterações heurística

+++ headers/Graph.h
@@ -35,8 +35,9 @@ public:
     vector<int> createHotelsCandidates();                                                          //~ Cria conjunto de candidatos de hoteis;
     vector<pair<int, double>> updateCandidates(vector<pair<int, double>> *candidates, Node *node); //~ Atualiza conjunto de candidatos;
     vector<vector<int>> createNeighborhood(vector<int> solution, vector<pair<int, double>> cand, vector<int> hotels);
-    int get_total_dist(vector<int> solution);
-    int get_total_points(vector<int> solution);
+    //vector<vector<int>> createNeighborhood(vector<int> solution, vector<int> hotels);
+    float get_total_dist(vector<int> solution);
+    float get_total_points(vector<int> solution);
     float get_ls_points(vector<int> solution);
     vector<vector<int>> heuristic();                                                             //~ Função heurística;
         vector<vector<int>> randomizedHeuristic(float alfa, int numIt, int seed);                    //~ Função heurística randomizada;
