+++ headers/Node.h
@@ -10,17 +10,17 @@ class Node {
         //* Atributes ---------------------------------------------------------------
         int id;            //~ Id do nó;
         char type;         //~ Tipo do nó (I = Hotel Inicial, F = Hotel Final, H = Hotel, N = Nó);
-        int points;        //~ Pontuação do nó (Se for um hotel, points = 0);
+        float points;        //~ Pontuação do nó (Se for um hotel, points = 0);
         
-        int pos_x;         //~ Posição X do nó;
-        int pos_y;         //~ Posição Y do nó;
+        float pos_x;         //~ Posição X do nó;
+        float pos_y;         //~ Posição Y do nó;
     
         Edge *first_edge;  //~ Primeira aresta;
         Edge *last_edge;   //~ Última aresta;
         Node *next_node;   //~ Próximo nó; 
 
         //* Functions ---------------------------------------------------------------
-        Node (int id, char type, int points, int x, int y);  //~ Constructor;
+        Node (int id, char type, float points, float x, float y);  //~ Constructor;
         ~Node ();                                            //~ Destructor;
 
         bool search_edge(int target_id);            //~ Verifica se a aresta já existe;
